requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-ollama>=0.1.0
ragas>=0.1.14
datasets>=2.19.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
greenlet>=3.0.0
pydantic-settings>=2.2.0
httpx>=0.27.0
pandas>=2.0.0

# Dev / test only
pytest>=8.0.0
pytest-asyncio>=0.23.0
